refector: 회원가입 폼 네이밍 명 변경 및 리팩토링
