+++ src/pages/sign-up/sign-up-agreement/form-agreement-modal.tsx
